#0: relocate test utils into ttnn/graph; add query apis
